requirements.txt
bcrypt==3.2.2
streamlit==1.23.0
streamlit-option-menu==0.3.2
SQLAlchemy==1.4.39
streamlit-pandas-profiling==0.1.3
seaborn==0.13.2
prophet==1.1.5 
plotly-express==0.4.1
openpyxl==3.0.10
PyJWT==2.4.0
#sentimentanalysis
numpy==1.26.2
pandas==2.1.4
#streamlit_pandas_profiling==0.1.2
#streamlit==1.11.0
scikit_learn==1.3.2
#plotly_express==0.4.1
plotly==5.1.0
wordcloud==1.8.1
nltk==3.6.2
textblob==0.15.3
findspark
pyspark
tweepy==3.10.0
python-decouple==3.4
python-dotenv==0.21.0
#wordcloud==1.8.1
# nltk==3.6.2
textblob==0.15.3
google-cloud-translate==2.0.1
googletrans==3.1.0a0
pydantic-settings
#attrs<20.0,>=19.1
holidays==0.25
altair==4.0
lightgbm
python-box
google-generativeai
#python-dotenv
pyarrow==14.0.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.112.0
#st-gsheets-connection
#sqlite3
